Bug: seqinr::s2c()
The seqinr::s2c() function takes a string as an argument and using the syntax aligned_sequences[1] may or may not return a single element of the list. I have encountered errors related to seqinr::s2c() receiving a list with a single element instead of a character. I have changed the syntax to [[]] (always returns a single list element).